Improved neighbour search, clean. Durations in s measured with clock function for LJHardBench (5000 timesteps): Old_Rc          6.81 New_Rc          13.47 New_Rc/2        9.04

+++ src/basic/source/data_format.cpp
@@ -87,7 +87,7 @@ DataFormat::attribute_t DataFormat::addAttribute
     if (s == "")
       s = name;
     attribute_t attr = {
-      name, m_attr_by_index.size(), m_size, datatype, persistent, s
+      name, static_cast<int> (m_attr_by_index.size()), m_size, datatype, persistent, s
     };
       
     m_size += c_size_of_datatype[datatype];
